refactored appartifact manager and cimaterialconfigservice
